Fixed Octree
The octree cells are no longer overlapping (this happened because the center was computed by each octree node itself and therefore the center was shifted). Now the parent node computes and assigns all the child node centers. Also the size was wrong (now it's half the length of an edge of the cube).